requirements.txt
fastapi
uvicorn
python-multipart
pillow
google-generativeai